Further ensure no repeats & add log message

+++ sbndcode/CRT/CRTUtils/CRTHitRecoAlg.h
@@ -14,7 +14,7 @@ namespace sbnd{
 
   struct CRTStripHit {
     CRTStripHit(uint32_t _channel, uint32_t _t0, uint32_t _t1, double _x,
-                double _ex, uint16_t _adc1, uint16_t _adc2, uint16_t _febdataindex) :
+                double _ex, uint16_t _adc1, uint16_t _adc2, uint8_t _febdataindex) :
       channel        (_channel)
       , t0           (_t0)
       , t1           (_t1)
@@ -32,7 +32,7 @@ namespace sbnd{
     double   ex;           // Error on lateral position [cm]
     uint16_t adc1;         // ADC 1st SiPM
     uint16_t adc2;         // ADC 2nd SiPM
-    uint16_t febdataindex; // Index of FEBData used to make strip hit
+    uint8_t  febdataindex; // Index of FEBData used to make strip hit
   };
